refractored into controllers and saved session token
